python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-telegram-bot==20.7
pandas==2.1.4
prophet==1.1.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
newsapi-python==0.2.7
python-dotenv==1.0.0
plotly==5.17.0
matplotlib==3.8.2